FIX: pages, components 하위 디렉토리명 확정
