+++ src/pages/auth/SignUp.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import useForm from '../../hooks/useForm';
-import Form from '../../components/auths/Form';
+import Form from '../../components/auth/Form';
 import { useAuthContext } from '../../context/AuthContext';
 
-export default function SignIn() {
+export default function SignUp() {
   const navigate = useNavigate();
   const auth = useAuthContext();
 
@@ -15,8 +15,8 @@ export default function SignIn() {
   const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
     try {
-      await auth.signin(values, () => {
-        navigate('/todo');
+      await auth.signup(values, () => {
+        navigate('/signin');
       });
     } catch (err) {
       console.error(err);
@@ -26,13 +26,13 @@ export default function SignIn() {
   return (
     <div className='bg-black-100 mt-[160px]'>
       <div className='w-[460px] my-0 mx-auto rounded-[16px] bg-white p-[20px]'>
-        <h1 className='mb-[24px] font-bold text-[24px]'>로그인</h1>
+        <h1 className='mb-[24px] font-bold text-[24px]'>회원가입</h1>
         <Form
-          handleSubmit={handleSubmit}
           values={values}
           validValues={validValues}
+          handleSubmit={handleSubmit}
           handleChange={handleChange}
-          pageName='signIn'
+          pageName='signUp'
         />
       </div>
     </div>
