+++ src/routes/Router.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Routes, Route, BrowserRouter, Navigate } from 'react-router-dom';
-import SignUp from '../pages/auths/SignUp';
-import SignIn from '../pages/auths/SignIn';
+import SignUp from '../pages/auth/SignUp';
+import SignIn from '../pages/auth/SignIn';
 import Todo from '../pages/todo/Todo';
-import PublicRoute from '../pages/authenticateds/PublicRoute';
-import PrivateRoute from '../pages/authenticateds/PrivateRoute';
+import PublicRoute from '../pages/authenticated/PublicRoute';
+import PrivateRoute from '../pages/authenticated/PrivateRoute';
 import { AuthProvider } from '../context/AuthContext';
 import { TodoProvider } from '../context/TodoContext';
